[DSCP-364] Split expanders into separate modules
