+++ witness/src/Dscp/Snowdrop/Expanders/Publication.hs
@@ -1,0 +1,102 @@
+-- | Publications expansion and validation.
+module Dscp.Snowdrop.Expanders.Publication
+    ( toProofPublicationTx
+    , seqExpandersPublicationTx
+    ) where
+
+import Data.Default (Default (..))
+import qualified Data.Map.Strict as Map
+import qualified Data.Set as Set
+
+import Snowdrop.Core (Expander (..), SeqExpanders (..), StateTxType (..), ValueOp (..), mkDiffCS,
+                      queryOne, queryOneExists)
+import Snowdrop.Util
+
+import Dscp.Core
+import Dscp.Crypto as Dscp
+import Dscp.Snowdrop.Configuration
+import qualified Dscp.Snowdrop.Configuration as Conf (Proofs (..))
+import Dscp.Snowdrop.Expanders.Common
+import Dscp.Snowdrop.Storage.Types
+import Dscp.Snowdrop.Types
+
+toProofPublicationTx :: Fees -> PublicationTxWitnessed -> (StateTxType, Proofs)
+toProofPublicationTx minFee (PublicationTxWitnessed tx (PublicationTxWitness {..})) =
+    (txType, Conf.PublicationTxWitnessProof $ WithSignature
+        { wsSignature = toDscpSig pwSig
+        , wsPublicKey = toDscpPK pwPk
+        , wsBody = (toPtxId tx, pwPk, ptHeader)
+        } `PersonalisedProof`
+            minFee
+    )
+  where
+    PublicationTx { ptHeader } = tx
+    txType = StateTxType $ getId (Proxy @TxIds) PublicationTxTypeId
+
+
+seqExpandersPublicationTx
+    :: Address
+    -> Fees
+    -> SeqExpanders Exceptions Ids Proofs Values ctx PublicationTxWitnessed
+seqExpandersPublicationTx feesReceiverAddr (Fees minFee) =
+    SeqExpanders $ one $ Expander
+        (Set.fromList [publicationOfPrefix])
+        (Set.fromList [accountPrefix, publicationHeadPrefix, publicationOfPrefix,
+                       publicationIdsPrefix, privateBlockTxPrefix])
+        $ \ptw@PublicationTxWitnessed { ptwTx } -> do
+            let PublicationTx{ ptHeader, ptAuthor, ptFeesAmount } = ptwTx
+            let ptxId = hash ptwTx
+            let phHash = hash ptHeader
+            let prevHash = ptHeader ^. pbhPrevBlock
+            let (prevHashM :: Maybe PrivateHeaderHash) =
+                    prevHash <$ guard (prevHash /= genesisHeaderHash ptAuthor)
+
+            headerWasEarlier <- queryOneExists (PublicationHead phHash)
+            let headerIsLast = prevHash == phHash
+            when (headerWasEarlier || headerIsLast) $
+                throwLocalError PublicationLocalLoop
+
+            let feeAmount = fromIntegral $ coinToInteger ptFeesAmount
+            maybePub <- queryOne (PublicationsOf ptAuthor)
+            mFeesReceiver <- queryOne (AccountId feesReceiverAddr)
+
+            let isPaid = ptAuthor /= feesReceiverAddr
+
+            account <- fromMaybe def <$> queryOne (AccountId ptAuthor)
+
+            when (isPaid && feeAmount < coinToInteger minFee) $
+                throwLocalError PublicationFeeIsTooLow
+                    { peMinimalFee = coinToInteger minFee, peGivenFee = feeAmount }
+
+            when (aBalance account < feeAmount) $
+                throwLocalError PublicationCantAffordFee
+                    { peFee = feeAmount, peBalance = aBalance account }
+
+            let dIssuer =
+                    maybe (New Account { aBalance = feeAmount, aNonce = 0 })
+                          (\issuer -> Upd issuer { aBalance = aBalance issuer + feeAmount })
+                          mFeesReceiver
+
+            let feesChanges
+                    | feeAmount == 0 = []
+                    | isPaid =
+                          [ AccountId ptAuthor ==>
+                                Upd account
+                                    -- You can say that `PubOf ==> LastPub`
+                                    -- mechanism will prevent double-pub
+                                    -- but lets not depend on that impl detail.
+                                    { aNonce   = aNonce account + 1
+                                    , aBalance = aBalance account - feeAmount
+                                    }
+                          , AccountId feesReceiverAddr ==> dIssuer ]
+                    -- If publication author and fees receiver are the same
+                    -- address, we do not transfer any coins.
+                    | otherwise = []
+
+            let change = if isJust maybePub then Upd else New
+            pure $ mkDiffCS $ Map.fromList $
+                [ PublicationsOf  ptAuthor ==> change (LastPublication phHash)
+                , PublicationHead phHash   ==> New    (PublicationNext prevHashM)
+                , PublicationIds  ptxId    ==> New    (PublicationItself ptw)
+                , PrivateBlockTx  phHash   ==> New    (PrivateBlockTxVal ptxId)
+                ] ++ feesChanges
